refactor: move test files into correct location
